fixed: todayscreen

+++ src/components/HabitList/index.js
@@ -4,6 +4,7 @@ import userContext from '../../Context/userContext'
 
 import { ToastContainer, toast } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
+import { FallingLines } from 'react-loader-spinner'
 
 import treshIcon from './../../assets/Vector(6).svg'
 import * as style from './style'
@@ -13,21 +14,31 @@ export default function HabitsList() {
     const { infoUser, habits, setHabits } = useContext(userContext);
     const [changeCollor, setChangeCollor] = useState(false)
     const [createNewHabit, setCreateNewHabit] = useState(false)
-    const [color,setColor]= useState('FFFFFF')
+    const [color, setColor] = useState('FFFFFF')
     const [newHabit, setNewHabit] = useState({
         name: undefined,
         days: [],
     })
-    let listWeekdayColor = ['#FFFFFF','#FFFFFF','#FFFFFF','#FFFFFF','#FFFFFF','#FFFFFF','#FFFFFF']
-    let listWeekday = newHabit.days
+    const [msgInput, setMsgInput] = useState('Salvar')
+    const loadingInput = <FallingLines width="45" color='#126BA5' />
+    const [isDisabled, setIsDisabled] = useState(false)
+    const [newHabitDetected, setnewHabitDetected] = useState(false)
+
     let week = ['D', 'S', 'T', 'Q', 'Q', 'S', 'S']
 
-    function selectWeekDay(item){
-        console.log(item.id)
-     
+    function selectWeekDay(item) {
+        if (item.className == 'selected') {
+            item.className = 'sc-breuTD iCNbXX'
+        } else {
+            item.className = 'selected'
+
+        }
+        console.log(item)
+
     }
 
     useEffect(() => {
+        console.log('entrou')
         axios({
             method: 'get',
             url: 'https://mock-api.bootcamp.respondeai.com.br/api/v2/trackit/habits',
@@ -38,38 +49,70 @@ export default function HabitsList() {
             setHabits(response.data)
             console.log(response.data)
         }).catch(err => console.log(err))
-    }, [habits])
+    }, [newHabitDetected])
+
+    function addNewDay(day) {
+        console.log(newHabit.days)
+        if (newHabit.days.includes(day)) {
+            const index = newHabit.days.indexOf(day);
+            if (index > -1) {
+                newHabit.days.splice(index, 1);
+            }
+        } else {
+          
+            setNewHabit({ ...newHabit, days: [...newHabit.days, day] })
+        }
+    }
 
     function sendNewHabit() {
+        setMsgInput(loadingInput)
+        setIsDisabled(true)
+        setnewHabitDetected(!newHabitDetected)
+
         axios({
             method: 'post',
             url: 'https://mock-api.bootcamp.respondeai.com.br/api/v2/trackit/habits',
             headers: {
                 "Authorization": `Bearer ${infoUser.token}`
             },
-            data: newHabit,
+            data: newHabit
         }).then(response => {
             console.log(response)
             toast.success('novo hábito adicionado!')
-            HabitsList()
+            setIsDisabled(false)
+            setNewHabit({name:'', days:[]})
+            setCreateNewHabit(false)
+            setMsgInput('Salvar')
+
+
         }).catch(err => {
             console.log(err)
             toast.warn('hábito não adicionado! por favor preencha todas as informações.')
-            console.log(infoUser, ' ', newHabit)
+            // console.log(infoUser, ' ', newHabit)
+            setIsDisabled(false)
+            setCreateNewHabit(false)
+            setNewHabit({name:'', days:[]})
+            setMsgInput('Salvar')
         })
 
     }
     function deletHabit(id) {
-        axios({
-            method: 'delete',
-            url: `https://mock-api.bootcamp.respondeai.com.br/api/v2/trackit/habits/${id}`,
-            headers: {
-                "Authorization": `Bearer ${infoUser.token}`
-            },
-        }).then(response => {
-            console.log(response)
-            toast.success('item deletado!')
-        }).catch(err => { console.log(err) })
+  
+        const deletitem = window.confirm('Tem certeza que deseja deletar esse hábito?')
+        if(deletitem){
+            setnewHabitDetected(!newHabitDetected)
+            axios({
+                method: 'delete',
+                url: `https://mock-api.bootcamp.respondeai.com.br/api/v2/trackit/habits/${id}`,
+                headers: {
+                    "Authorization": `Bearer ${infoUser.token}`
+                },
+            }).then(response => {
+                console.log(response)
+                toast.success('item deletado!')
+                setnewHabitDetected(!newHabitDetected)
+            }).catch(err => { console.log(err) })
+        }
     }
 
 
@@ -80,7 +123,7 @@ export default function HabitsList() {
                 <h1>Meus hábitos</h1>
                 <div className="containerImg" onClick={() => {
                     setCreateNewHabit(true)
-                    console.log('click')
+
                 }}>
                     <img src={icon} alt='' />
                 </div>
@@ -90,19 +133,21 @@ export default function HabitsList() {
                     <style.NewHabit>
                         <input type='text' placeholder="nome do hábito" onChange={(e) => {
                             setNewHabit({ ...newHabit, name: e.target.value })
-                            console.log(newHabit)
-                        }} />
+
+                        }} disabled={isDisabled} />
                         <style.Days>
                             {week.map((item, index) => {
                                 return (
                                     <style.Day
-                                        key={index}
+                                        key={index} name={false}
                                         onClick={(e) => {
-                                            setNewHabit({ ...newHabit, days: [...newHabit.days, e.target.id] })
+                                            addNewDay(e.target.id)
+                                            
+                                            // setNewHabit({ ...newHabit, days: [...newHabit.days, e.target.id] })
                                             selectWeekDay(e.target)
                                         }}
                                         id={index}
-                                        bg={changeCollor ? '#126BA5' : 'FFFFFF'}>{item}
+                                        bg={'FFFFFF'}>{item}
                                     </style.Day>
                                 )
                             })}
@@ -113,18 +158,17 @@ export default function HabitsList() {
                                 toast.error('cancelado!')
                                 setCreateNewHabit(false)
                             }
-                            } Bgcolor='none' font='#52B6FF;' positionBotton='15px' positionRight='123px'>
+                            } Bgcolor='none' font='#52B6FF;' positionBotton='15px' positionRight='123px' disabled={isDisabled}>
                                 Cancelar
                             </style.ButtonNewHabit>
 
                             <style.ButtonNewHabit onClick={() => {
                                 sendNewHabit()
 
-                                setCreateNewHabit(false)
 
-                            }} Bgcolor='#52B6FF;' font='#FFFFFF;' positionBotton='15px' positionRight='16px'>
-                                Salvar
-                            </style.ButtonNewHabit>
+                            }} Bgcolor='#52B6FF;' font='#FFFFFF;' positionBotton='15px' positionRight='16px' disabled={isDisabled}>
+                                {msgInput}
+                            </style.ButtonNewHabit >
                         </style.ButtonsNewHabit>
                     </style.NewHabit> : <></>
                 }
